test: Parser performance test script
perfscript is run with name of a pcap file and the name of a core as in

	./perfscript test.pcap panda

It computes a compute a performance estimate for the core for the
workload in the pcap. It will determine a -n value which approximates
one hundred million packets total, based on the number of packets in the
pcap file.

Signed-off-by: Jamal Hadi Salim <hadi@mojatatu.com>