tlvs: TLV parsing in the PANDA parser
Definitions and parsing engine for TLVs in the PANDA parser.

parser.h contains structure definition and helper functions for TLV
nodes, and tables.

parser.c implements adds TLV handling in the main parsing loop for
those parse nodes (and protocol nodes) that contain TLVs. For a parse
node that contains TLVs (a TLVs parse node) the TLVs are processed
after extracting metadata from the protocol layer. TLV parsing occurs
in a loop in the panda_parser_tlvs functions. To parse the TLVs, the
TLV length and type functions in the TLVs parse node are invoked. For
each TLV the extract metadata and process TLV functions of individual
TLV parse nodes may be invoked.

Signed-off-by: Tom Herbert <tom@sipanda.io>