samples: Add support to sample parsers to use optimized parser
Allow the two sample parsers to use the optimized compiler than is
produced by the panda-compiler tool. Change the Makefile to build
the opitimized parser and non-optimized one as well. Add -O argument
to parser_tmpl and parser_notmpl which when set indicates to use the
optimized parser else use the non-optimized parser.

Signed-off-by: Tom Herbert <tom@sipanda.io>