xdp: TCP options hack
In order to support variable length TLVs with metadata extraction,
specifically TCP option, explict eBPF functions for delaing with the
common TCP options are defined (in panda/bpf.h). The panda-compiler
generate code that calls the functions. Note this is a temporary hack
until a better solution is found to deal with variable length TLVs.