panda-compiler: Move metadata_all out from panda-compiler
When XDP support was introduced in the panda-compiler references
were made to panda_metadata_all in the compiler, but that is
just one metadata structure that shouldn't be hardcoded into the
compiler. struct panda_ctx is also defined by the compiler to
contain that metadata structures for XDP.

Remove struct panda_ctx from the compiler and references to
panda_metadata_all. A new struct panda_ctx is created in
include/panda/parser.h. That structure can be used as the basis
for a per application context defition where the user defined
structure contains a struct panda_ctx followed by metadata
frames (such as of type struct panda_metadata_all).