parser: Add PANDA_PARSER_XDP and PANDA_PARSE_XDP
This is used to define a PANDA Parser for XDP. Verifier does not
like calling the parser function via indirect entry points in the
parser, so PANDA_PARSE_XDP is added as a stop gap. This presumes
that the panda-compiler, or user, has created a any entry point
function for a parser as panda_xdp_parser_##PARSER where PARSER
if the name of the struct panda_parser. For defining an XDP parser
to be visible to the panda-compiler, PANDA_PARSER macro is
sufficient (as opposed to PANDA_PARSER_ADD which was being used.