xdp: Add template program for PANDA Parser in XDP
Create panda/xdp_tmpl.h that contains a full template for a XDP
program that invokes the PANDA parser.

The macro PANDA_XDP_MAKE_PARSER_PROGRAM is invoked to create an XDP
program.  The macro takes four arguments. The first is the name of
parser (i.e. a struct of type struct panda_parser).

The second argument is a structure type containing the calling context
for the XDP program. The first element of this structure must be a
struct panda_ctx and be named 'ctx'. The rest of the elements contain
the metadata frames of a custom structure type for the application.

The third argument is the metadata frame size. For instance, if the
canned metadata_all is being used then this argument could be set to
sizeof(struct panda_metadata_all).

The fourth argument is the processing function that is called after
the packet has been parsed. This function takes two arguments the
struct xdp_md and struct panda_ctx for the packet. The processing
function returns and XDP disposition code (e.g. XDP_PASS, XDP_DROP,
etc.).

The fifth argument is an error handler function that is called after
when the parser fails. This function takes three arguments: a PANDA
error code (i.e. PANDA_* error codes), the struct xdp_md and struct
panda_ctx for the packet. The error handler function returns and XDP
disposition code (e.g. XDP_PASS, XDP_DROP, etc.).