parser: TLV overlay parse nodes
Support for TLV overlay parse nodes. This is helpful to deal with
variable length TLVs such as TCP sacks. The overlay allows a secondary
level of parsing for TLVs for some type. In the case of TCP sacks, an
overlay is matched to the option length which can be one of four values
for the sack option (10, 18, 26, or 34). With the length matched the
processing and metadata functions can operate on the variant of sack
options using fixed fields. For instance, if a sack option length is
18 then the metadata extraction can directly access the sack array
at [0] and [1]. By creating a overlay for each valid option length,
the use of a loop in the metadata extraction is eschewed.

This patch includes:
    * Add overlay_table to struct panda_parse_tlv_node. Type is a
      struct panda_proto_tlvs_table. This indicates an overlay table
      for a TLV node. If it is set to NULL then the TLV node is not an
      overlay node
    * Add overlay_type to struct panda_parse_tlv_node_ops. This extracts
      the sub-TLV type to lookup in an overlay table. If this is not set
      then the TLV length is used as the type by default
    * Add unknown_overlay_type to struct panda_parse_tlv_node_ops. This
      is called if no TLV node is found in an overlay lookup. If the
      function returns PANDA_OKAY then the TLV is skipped, else parsing
      stops and the PANDA error is returned
    * Add helper macro PANDA_MAKE_TLV_OVERLAY_PARSE_NODE to create a
      TLV overlay node
    * In parser.c create to panda_parse_one_tlv function to parse one
      TLV
    * After extracting TLV metadata and TLV handling check if an
      overlay table is present in the TLV parse node. If it is, do a
      lookup in the table (type determined by overlay_type function
      or default using TLV length)
    * If an overlay is found process the TLV (got beginning of function
      to operate on next node)
    * Note that the code allows for nested overlay nodes