parser_big: Use TLV overlay node for TCP sacks
Create metadata functions for one to four sacks and TLV nodes for
one to four sacks that include the respective metadata extrction
function. Create a TLV overlay node for TCP sacks using
PANDA_MAKE_TLV_PARSE_NODE where the type for lookup in the overlay
table is the TLV length. The overlay contains four entries with
types 10, 18, 26, and 34 that respectively map to the TLV nodes
for one to four sacks.