parser: Debugging for TLVS and flag-fields
Add a name field to panda_parse_flag_field_node and
panda_parse_tlv_node. In the helper macros for creating TLV and
flag-field parse nodes, the name is set as the string of the node
name. When debugging is set (PANDA_F_DEBUG is set in input flags
to the parser function), print a debug message for parsing the
flag-field or TLV.