parser: Change unknown overlay TLV handling to use wildcards
This patch eliminates the TLV parse node callback functions
unknown_overlay_type. A wildcard node is added to TLV parse nodes that
is used as the overlay TLV node when a lookup fails to find a hit. The
wildcard node can be a catchall for default TLV processing, or could be
an error handling node (i.e. its handler function can return a PANDA
parser error.

Additionally, each parse node contains and "unknown_overlayret" field.
This is a PANDA parser code that is returned when a TLV overlay lookup
fails and the wildcard node is false.