parser: Change unknown next proto handling to use wildcard nodes
This patch eliminates the callback function unknown_next_proto (as well
as panda_unknown_next_proto_fail and panda_unknown_next_proto_ignore).
A wildcard node is added to parse noded that is used as the next
protocol node when a lookup fails to find a hit. The wildcard node
can be a catchal for other protocols not in the table, or could be
an error handling node (i.e. its handler function can return a
PANDA parser error).

Additionally, each parse node contains and "unknown_ret" field. This
is a PANDA parser code that is returned when a lookup fails and the
wildcard node is set to NULL.