parser: Change unknown TLV handling to use wildcard nodes
This patch eliminates the TLVs parse node callback function unknown_type
(as well as panda_unknown_tlv_fail and panda_unknown_tlv_ignore). A
wildcard node is added to parse TLV nodes that is used as the TLV node
when a TLV lookup fails to find a node. The wildcard node would
typically be an error handling node (i.e. its handler function can
return a PANDA parser error).

Additionally, each parse TLVs node contains and "unknown_tlv_type_ret"
field.  This is a PANDA parser code that is returned when a TLV lookup
fails and the wildcard node is false.