parser: Add parser tables
Parser tables allow mapping a value, probably a protocol number, to
a parser. This is useful for instance to arbitrarily parse a packet
starting at some known L3 protocol (where the protocol number is
external input to the parsing and not contained in the packet
headers itself).

A parser table is created by PANDA_MAKE_PARSER_TABLE and a table
is constructed of an array of structures containing a key value
and a pointer pointer (**) to a parser (double pointer is used since
parsers may not be defined as constants, see PANDA_PARSER_ADD).

The API function panda_parse_from_table is added to parse a packet
based on a protocol table, the function parameters includes a
pointer to a protocol table and the key value to match.

An effect of this feature is that parsing from node is no longer
needed, the same thing can be accomplished by create different parsers
for the different start nodes being used. Accordingly, the start node
argument in __panda_parse and panda_parser_opt_entry_point is removed.